requirements.txt
Django==1.8.1
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.3.0
psycopg2==2.6
static==0.4
wsgiref==0.1.2
djangorestframework